추천 수 기능 추가

+++ backend/community/src/main/java/com/teamhs/community/domain/Board.java
@@ -18,7 +18,7 @@ public class Board {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // Auto-increment strategy
-    private int boardId;
+    private Long boardId;
 
     @Column(nullable = true)
     private String userId;
